requirements.txt
argcomplete==3.5.1
asgiref==3.8.1
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
durationpy==0.9
et_xmlfile==2.0.0
gunicorn==23.0.0
hjson==3.1.0
idna==3.10
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
placebo==0.9.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.0
troposphere==4.8.3
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.0
XlsxWriter==3.2.0
zappa==0.58.0
